Fix runtime crash when power columns are entirely null
- Add explicit cast to `pl.Float32` before calling `.abs()` in `calculate_target_map` to prevent `InvalidOperationError` on Null dtype columns.
- Add `test_calculate_target_map_all_null_power` to `packages/ml_core/tests/test_target_selection.py` to verify graceful handling of missing data.
- Verified fix with pytest and ty check.

+++ packages/ml_core/src/ml_core/data.py
@@ -46,7 +46,8 @@ def calculate_target_map(
         mva_last_seen=pl.col("timestamp").filter(pl.col(POWER_MVA).is_not_null()).max(),
         max_timestamp=pl.col("timestamp").max(),
         peak_capacity_MW_or_MVA=pl.max_horizontal(
-            pl.col(POWER_MW).abs().max(), pl.col(POWER_MVA).abs().max()
+            pl.col(POWER_MW).cast(pl.Float32).abs().max(),
+            pl.col(POWER_MVA).cast(pl.Float32).abs().max(),
         ).fill_null(1.0),
     )
